rename top level packagename to fred
