+++ src/fred/api/categories.py
