+++ src/fred/api/exceptions.py
