+++ src/fred/api/series.py
