+++ src/fred/util/decorators.py
